fix: implement all CRITICAL/HIGH/MEDIUM/LOW bugfixes from implementation plan
- Architecture: unified snake_case→camelCase transformation via transformKeys()
- CRITICAL-1: TierDriftTracker now uses canonical CurrencyTier/TiersResponse types
- CRITICAL-2: Fixed league default mismatch (vaal→runes) across all files
- CRITICAL-3: All fallback data converted to camelCase
- HIGH-1: spread display uses spread with fallback
- HIGH-2: Deduplicated scoreColor function
- HIGH-3: Fixed pct?.toFixed(1) rendering +undefined%
- HIGH-4: Deleted dead auth code (routes_auth.py, auth/start, auth/callback)
- HIGH-5: Eliminated double _compute_cross_rate_divergence call (TriangularResult)
- HIGH-6: PricePoint.volume type int→float
- HIGH-7: Renamed confidence_level→significance_level
- HIGH-8: Deleted dead test_gold_costs.py
- MEDIUM-1..13: config sync, caching, timestamp heuristic, dead code, etc.
- LOW-1..12: binary search, BFS docs, test improvements, etc.

+++ backend/models/currency.py
@@ -68,7 +68,7 @@ class PricePoint:
     """A single price observation at a point in time."""
     timestamp: datetime
     price: float           # price in base currency (Exalted)
-    volume: int = 0
+    volume: float = 0.0
 
 
 @dataclass(frozen=True)
